+++ src/odr/internal/html/image_file.hpp
@@ -14,9 +14,9 @@ class Logger;
 namespace odr::internal::html {
 
 void translate_image_src(const File &file, std::ostream &out,
-                         const HtmlConfig &config);
+                         const HtmlConfig &config, const Logger &logger);
 void translate_image_src(const ImageFile &image_file, std::ostream &out,
-                         const HtmlConfig &config);
+                         const HtmlConfig &config, const Logger &logger);
 
 HtmlService create_image_service(const ImageFile &image_file, HtmlConfig config,
                                  const Logger &logger);